pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["Atlas_v2"]

[project]
name = "axis-ai"
version = "2.7.21"
description = "Autonomous Spatial AI & MCP Framework"
authors = [
    {name = "Eugen1189", email = "your.email@example.com"}
]
dependencies = [
    "google-generativeai==0.8.5",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "urllib3<2.0.0",
    "pyyaml==6.0.3",
    "psutil>=5.9.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "numpy>=2.3.0,<3.0.0",
    "opencv-python>=4.10.0",
    "mediapipe>=0.10.0",
    "mcp>=1.1.2",
    "googlesearch-python>=1.2.3",
    "structlog>=24.4.0",
    "ollama>=0.6.1",
    "PyQt6>=6.6.0",
    "chromadb>=0.5.0",
    "speechrecognition>=3.10.0",
    "pyttsx3>=2.90",
    "nest-asyncio>=1.6.0",
    "Pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "pygame>=2.5.0"
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.3.1",
    "vulture>=2.14",
    "deptry>=0.23.0",
    "ruff>=0.3.0",
    "google-genai==1.55.0"
]

[tool.deptry]
pep621_dev_dependency_groups = ["dev"]
package_module_name_map = { "nest-asyncio" = "nest_asyncio" }
per_rule_ignores = { DEP002 = ["urllib3", "pytest-asyncio", "pytest-cov", "pytest-timeout", "vulture", "deptry", "ruff"], DEP004 = ["pytest"] }



[tool.pytest.ini_options]

pythonpath = ["Atlas_v2"]
testpaths = ["Atlas_v2/tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
